move spark logging
