+++ pom.xml
@@ -44,7 +44,7 @@
         <akka.version>2.4.9</akka.version>
         <json4s.version>3.4.0</json4s.version>
         <curator.version>3.2.0</curator.version>
-        <spark.version>2.0.0</spark.version>
+        <spark.version>2.0.2</spark.version>
         <slf4j.version>1.7.7</slf4j.version>
         <scalaz-version>7.2.1</scalaz-version>
         <!-- Scala version and cross build properties -->
